rewrite opnode to store usize instead of Op
