+++ rust/automerge/src/query/keys_at.rs
@@ -1,4 +1,4 @@
-use crate::op_tree::OpTreeNode;
+use crate::op_tree::OpTreeInternal;
 use crate::query::VisWindow;
 use crate::types::{Clock, Key};
 use std::fmt::Debug;
@@ -11,19 +11,19 @@ pub(crate) struct KeysAt<'a> {
     last_key: Option<Key>,
     index_back: usize,
     last_key_back: Option<Key>,
-    root_child: &'a OpTreeNode,
+    op_tree: &'a OpTreeInternal,
 }
 
 impl<'a> KeysAt<'a> {
-    pub(crate) fn new(root_child: &'a OpTreeNode, clock: Clock) -> Self {
+    pub(crate) fn new(op_tree: &'a OpTreeInternal, clock: Clock) -> Self {
         Self {
             clock,
             window: VisWindow::default(),
             index: 0,
             last_key: None,
-            index_back: root_child.len(),
+            index_back: op_tree.len(),
             last_key_back: None,
-            root_child,
+            op_tree,
         }
     }
 }
@@ -33,7 +33,7 @@ impl<'a> Iterator for KeysAt<'a> {
 
     fn next(&mut self) -> Option<Self::Item> {
         for i in self.index..self.index_back {
-            let op = self.root_child.get(i)?;
+            let op = self.op_tree.get(i)?;
             let visible = self.window.visible_at(op, i, &self.clock);
             self.index += 1;
             if Some(op.elemid_or_key()) != self.last_key && visible {
@@ -48,7 +48,7 @@ impl<'a> Iterator for KeysAt<'a> {
 impl<'a> DoubleEndedIterator for KeysAt<'a> {
     fn next_back(&mut self) -> Option<Self::Item> {
         for i in self.index..self.index_back {
-            let op = self.root_child.get(i)?;
+            let op = self.op_tree.get(i)?;
             let visible = self.window.visible_at(op, i, &self.clock);
             self.index_back -= 1;
             if Some(op.elemid_or_key()) != self.last_key_back && visible {
