+++ rust/automerge/src/query/list_range.rs
@@ -1,5 +1,5 @@
 use crate::exid::ExId;
-use crate::op_tree::OpTreeNode;
+use crate::op_tree::OpTreeInternal;
 use crate::types::{ElemId, OpId};
 use crate::values::ValueIter;
 use crate::{Automerge, Value};
@@ -14,19 +14,19 @@ pub(crate) struct ListRange<'a, R: RangeBounds<usize>> {
     last_elemid: Option<ElemId>,
     next_result: Option<(usize, Value<'a>, OpId)>,
     index_back: usize,
-    root_child: &'a OpTreeNode,
+    op_tree: &'a OpTreeInternal,
 }
 
 impl<'a, R: RangeBounds<usize>> ListRange<'a, R> {
-    pub(crate) fn new(range: R, root_child: &'a OpTreeNode) -> Self {
+    pub(crate) fn new(range: R, op_tree: &'a OpTreeInternal) -> Self {
         Self {
             range,
             index: 0, // FIXME root_child.seek_to_pos(range.start)
             pos: 0,   // FIXME range.start
             last_elemid: None,
             next_result: None,
-            index_back: root_child.len(),
-            root_child,
+            index_back: op_tree.len(),
+            op_tree,
         }
     }
 }
@@ -45,7 +45,7 @@ impl<'a, R: RangeBounds<usize>> Iterator for ListRange<'a, R> {
     // point and stop at the end point and not needless scan all the ops before and after the range
     fn next(&mut self) -> Option<Self::Item> {
         for i in self.index..self.index_back {
-            let op = self.root_child.get(i)?;
+            let op = self.op_tree.get(i)?;
             self.index += 1;
             if op.visible() {
                 if op.elemid() != self.last_elemid {
