+++ rust/automerge/src/query/list_range_at.rs
@@ -1,6 +1,6 @@
 use super::VisWindow;
 use crate::exid::ExId;
-use crate::op_tree::OpTreeNode;
+use crate::op_tree::OpTreeInternal;
 use crate::types::{Clock, ElemId, OpId};
 use crate::values::ValueIter;
 use crate::{Automerge, Value};
@@ -15,7 +15,7 @@ pub(crate) struct ListRangeAt<'a, R: RangeBounds<usize>> {
     last_elemid: Option<ElemId>,
     next_result: Option<(usize, Value<'a>, OpId)>,
     index_back: usize,
-    root_child: &'a OpTreeNode,
+    op_tree: &'a OpTreeInternal,
     clock: Clock,
     window: VisWindow,
 }
@@ -27,15 +27,15 @@ impl<'a, R: RangeBounds<usize>> ValueIter<'a> for ListRangeAt<'a, R> {
 }
 
 impl<'a, R: RangeBounds<usize>> ListRangeAt<'a, R> {
-    pub(crate) fn new(range: R, clock: Clock, root_child: &'a OpTreeNode) -> Self {
+    pub(crate) fn new(range: R, clock: Clock, op_tree: &'a OpTreeInternal) -> Self {
         Self {
             range,
             index: 0, // FIXME root_child.seek_to_pos(range.start)
             pos: 0,   // FIXME range.start
             last_elemid: None,
             next_result: None,
-            index_back: root_child.len(),
-            root_child,
+            index_back: op_tree.len(),
+            op_tree,
             clock,
             window: VisWindow::default(),
         }
@@ -47,7 +47,7 @@ impl<'a, R: RangeBounds<usize>> Iterator for ListRangeAt<'a, R> {
 
     fn next(&mut self) -> Option<Self::Item> {
         for i in self.index..self.index_back {
-            let op = self.root_child.get(i)?;
+            let op = self.op_tree.get(i)?;
             let visible = self.window.visible_at(op, i, &self.clock);
             self.index += 1;
             if visible {
