+++ rust/automerge/src/query/opid.rs
@@ -33,7 +33,7 @@ impl OpIdSearch {
 }
 
 impl<'a> TreeQuery<'a> for OpIdSearch {
-    fn query_node(&mut self, child: &OpTreeNode) -> QueryResult {
+    fn query_node(&mut self, child: &OpTreeNode, _ops: &[Op]) -> QueryResult {
         if child.index.ops.contains(&self.target) {
             QueryResult::Descend
         } else {
